refactor(ux): isolate project status presentation
